+++ allauth/socialaccount/tests.py
@@ -21,7 +21,7 @@ import allauth.app_settings
 from ..account import app_settings as account_settings
 from ..account.models import EmailAddress
 from ..account.utils import user_email, user_username
-from ..tests import MockedResponse, TestCase, mocked_response
+from ..tests import Mock, MockedResponse, TestCase, mocked_response, patch
 from ..utils import get_user_model
 from . import app_settings, providers
 from .helpers import complete_social_login
@@ -301,7 +301,7 @@ class OAuth2TestsMixin(object):
         """
         self.test_account_tokens(multiple_login=True)
 
-    def login(self, resp_mock, process="login", with_refresh_token=True):
+    def login(self, resp_mock=None, process="login", with_refresh_token=True):
         resp = self.client.post(
             reverse(self.provider.id + "_login")
             + "?"
@@ -383,6 +383,54 @@ def create_oauth2_tests(provider):
 
     Class.__name__ = "OAuth2Tests_" + provider.id
     return Class
+
+
+class OpenIDConnectTests(OAuth2TestsMixin):
+    oidc_info_content = {
+        "authorization_endpoint": "/login",
+        "userinfo_endpoint": "/userinfo",
+        "token_endpoint": "/token",
+    }
+    userinfo_content = {
+        "picture": "https://secure.gravatar.com/avatar/123",
+        "email": "ness@some.oidc.server.onett.example",
+        "sub": 2187,
+        "identities": [],
+        "name": "Ness",
+    }
+    extra_data = {
+        "picture": "https://secure.gravatar.com/avatar/123",
+        "email": "ness@some.oidc.server.onett.example",
+        "sub": 2187,
+        "identities": [],
+        "name": "Ness",
+    }
+
+    def setUp(self):
+        super(OpenIDConnectTests, self).setUp()
+        patcher = patch(
+            "allauth.socialaccount.providers.openid_connect.views.requests",
+            get=Mock(side_effect=self._mocked_responses),
+        )
+        self.mock_requests = patcher.start()
+        self.addCleanup(patcher.stop)
+
+    def get_mocked_response(self):
+        # Enable test_login in OAuth2TestsMixin, but this response mock is unused
+        return True
+
+    def _mocked_responses(self, url, *args, **kwargs):
+        if url.endswith("/.well-known/openid-configuration"):
+            return MockedResponse(200, json.dumps(self.oidc_info_content))
+        elif url.endswith("/userinfo"):
+            return MockedResponse(200, json.dumps(self.userinfo_content))
+
+    @override_settings(SOCIALACCOUNT_AUTO_SIGNUP=True)
+    def test_login_auto_signup(self):
+        resp = self.login()
+        self.assertRedirects(resp, "/accounts/profile/", fetch_redirect_response=False)
+        sa = SocialAccount.objects.get(provider=self.provider.id)
+        self.assertDictEqual(sa.extra_data, self.extra_data)
 
 
 class SocialAccountTests(TestCase):