+++ test_settings.py
@@ -127,6 +127,7 @@ INSTALLED_APPS = (
     "allauth.socialaccount.providers.nextcloud",
     "allauth.socialaccount.providers.odnoklassniki",
     "allauth.socialaccount.providers.openid",
+    "allauth.socialaccount.providers.openid_connect",
     "allauth.socialaccount.providers.openstreetmap",
     "allauth.socialaccount.providers.orcid",
     "allauth.socialaccount.providers.patreon",
@@ -206,3 +207,21 @@ PASSWORD_HASHERS = [
 
 ACCOUNT_EMAIL_CONFIRMATION_COOLDOWN = 0
 ACCOUNT_RATE_LIMITS = {}
+
+
+SOCIALACCOUNT_PROVIDERS = {
+    "openid_connect": {
+        "SERVERS": [
+            {
+                "id": "unittest-server",
+                "name": "Unittest Server",
+                "server_url": "https://unittest.example.com",
+            },
+            {
+                "id": "other-server",
+                "name": "Other Example Server",
+                "server_url": "https://other.example.com",
+            },
+        ],
+    }
+}