+++ ChangeLog.rst
@@ -6,10 +6,11 @@ Note worthy changes
 
 - Officially support Django 4.1.
 
+- New providers: OpenID Connect, Twitter (OAuth2).
+
 - Introduced a new provider setting ``OAUTH_PKCE_ENABLED`` that enables the
   PKCE-enhanced Authorization Code Flow for OAuth 2.0 providers.
 
-- New providers: Twitter (OAuth2).
 
 
 0.51.0 (2022-06-07)